[KAFKA][YCQL][YSQL] YugabyteDB Sink for Kafka Connect #7586
Summary:
YugabyteDB Sink for Kafka Connect supporting both the YCQL as well as YSQL APIs.

The Sink should be able to take advantage of linear scaling and the high availability feature of YugabyteDB. Below are some of the requirements of the Sink.

We support adding multiple endpoints for the Sink which are distributed across the tasks randomly.

Test Plan:
Have added JUnit tests based on the TestContainer framework.
There have been benchmark tests also run for the Sinks.

Reviewers: hbhanawat, nkumar, nikhil, zyu

Reviewed By: zyu

Differential Revision: https://phabricator.dev.yugabyte.com/D10867