Removed repository related to Datastax for dependency
Summary:
A user claimed that it DataStax repository is not accessible for maven build. Although it is accessible from AWS machines.
The same jars are available in confluent repository too so removing the DataStax repository as they are not needed anyway.

Test Plan: Did a build on a fresh VM to verify if the connector jar is build successfully.

Reviewers: zyu

Reviewed By: zyu

Differential Revision: https://phabricator.dev.yugabyte.com/D11588